Add token based authentication
The authentication is based on a symmetric encryption
algorithm.

Tokens contains all the needed infromation to
authenticate the user on each incomming request.
That means tokens are not needed to be stored in a
backend system.

A token remains valid as long as the expiration date
is not met. There's no way to invalidate tokens.

In this first implementation token are no longer
valid after a server restart as long as a new secret
key is generated at server startup and never stored.