Fix blade IP resolution
IP address resolution was working only when the DHCP server
was running within the RuggedPOD controller. Few weeks ago we
introduced the DHCP-Proxy mode to leverage situations where
DHCP is running elsewhere on the network. In this case, reading
the DHCP lease file on the RuggedPOD controller makes no sense.

Now in this case, we read the IP address from the ARP table. If the
IP is not found, we try a discovery using nmap to populate the
ARP table and then retry to find the MAC address in the table.

There is still a problem with this approach. If the network range
is too big, for instance a /16 CIDR, the discovery is very long
(several minutes). For now, we recommend to use /24 networks or
smaller.

This last problem will be addressed later.