Add an API to execute commands on blades
It consists in the following endpoints:

* POST /blades/<blade_id>/commands
* GET  /blades/<blade_id>/commands/<id>
* GET  /blades/<blade_id>/commands
* GET  /commands/<id>
* GET  /commands

Commands are stored in the database, first to handle their lifecycle,
then to keep track of the past commands. At any time a command can
be requested to get information like its status (PENDING, SUCCESS or
ERROR) or any other information stored in the database.

Currently, commands are always executed through SSH. Nevertheless,
the API design does not lock-in to SSH. We could for example, run
commands through the serial port.

Here is an example of a command submission

```
curl -H "Content-Type: application/json" \
     -H "X-Auth-Token: ..." -X POST -d '
{
    "method": "ssh",
    "command": "sudo dmidecode",
    "username": "ruggedpod",
    "private_key": "..."
}' \
http://ruggedpod/api/v2/blades/3/commands
```